Move intent and associated types back to request namespace. Update Dialog and tests to use request namespace
